# gestimowebfront-end/src/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API pour GESTIMOWEB REST API",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger GESTIMOWEB REST API",
    "contact" : {
      "name" : "bzadata",
      "email" : "michel.bossoh@outlook.fr"
    }
  },
  "basePath" : "http://localhost:8282/",
  "paths" : {
...
    "gestimoweb/api/v1/image/imagesbybien/{id}" : {
      "get" : {
        "operationId" : "telechagerImage",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ImageDataDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/image/upload/{id}/{name}" : {
      "post" : {
        "summary" : "Téléchargder une image",
        "description" : "",
        "operationId" : "uploadImage",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "file",
          "in" : "query",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/affichetoutlesimmeubles/{idAgence}" : {
      "get" : {
        "operationId" : "affichageDesImmeubles",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ImmeubleEtageDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/all/{idAgence}" : {
      "get" : {
        "operationId" : "findAllImmeuble",
        "parameters" : [ {
          "name" : "idAgence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ImmeubleDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/deleteImmeuble/{id}" : {
      "delete" : {
        "operationId" : "deleteImmeuble",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/findById/{id}" : {
      "get" : {
        "operationId" : "findImmeubleByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ImmeubleDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/findByIdSite/{id}" : {
      "get" : {
        "operationId" : "findImmeubleByIdSite",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ImmeubleDto"
              }
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/findByName/{name}" : {
      "get" : {
        "operationId" : "findImmeubleByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ImmeubleDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/immeuble/saveImeubleEtage" : {
      "post" : {
        "operationId" : "saveImmeubleEtage",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ImmeubleEtageDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ImmeubleEtageDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/pays/delete/{id}" : {
      "delete" : {
        "operationId" : "deletePays",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/pays/findById/{id}" : {
      "get" : {
        "operationId" : "findPaysByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PaysDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/pays/findByName/{name}" : {
      "get" : {
        "operationId" : "findPaysByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PaysDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/pays/save" : {
      "post" : {
        "operationId" : "savePays",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PaysDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PaysDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/prestation/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteServiceAdditionnelPrestation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "gestimoweb/api/v1/prestation/findById/{id}" : {
      "get" : {
        "operationId" : "findServiceAdditionnelByIDPrestation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/prestation/saveorupdate" : {
      "post" : {
        "operationId" : "saveorupdatePrestation",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationSaveOrUpdateDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/reservation/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteReservation",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "gestimoweb/api/v1/reservation/saveorupdate" : {
      "post" : {
        "operationId" : "saveorupdateRes",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ReservationSaveOrUpdateDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ReservationSaveOrUpdateDto"
            }
          }
        }
      }
    },
...
    "gestimoweb/api/v1/serviceadditionnel/delete/{id}" : {
      "delete" : {
        "operationId" : "deleteServiceAdditionnelPrestationAdditionnel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "gestimoweb/api/v1/serviceadditionnel/findById/{id}" : {
      "get" : {
        "operationId" : "findServiceAdditionnelByIDPrestationAdditionnel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationAdditionnelReservationSaveOrrUpdate"
            }
          }
        }
      }
    },
    "gestimoweb/api/v1/serviceadditionnel/saveorupdate" : {
      "post" : {
        "operationId" : "saveorupdatePrestationAdditionnel",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PrestationAdditionnelReservationSaveOrrUpdate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PrestationAdditionnelReservationSaveOrrUpdate"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ImageDataDto" : {
      "type" : "object",
      "properties" : {
        "nameImage" : {
          "type" : "string"
        },
        "typeImage" : {
          "type" : "string"
        },
        "profileAgenceImageUrl" : {
          "type" : "string"
        },
        "imageData" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "byte"
          }
        },
        "bienimmobilier" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "ImmeubleDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeNomAbrvImmeuble" : {
          "type" : "string"
        },
        "nomCompletImmeuble" : {
          "type" : "string"
        },
        "nomBaptiserImmeuble" : {
          "type" : "string"
        },
        "descriptionImmeuble" : {
          "type" : "string"
        },
        "numImmeuble" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrEtage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrePiecesDansImmeuble" : {
          "type" : "integer",
          "format" : "int32"
        },
        "idSite" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "garrage" : {
          "type" : "boolean"
        }
      }
    },
    "ImmeubleEtageDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeNomAbrvImmeuble" : {
          "type" : "string"
        },
        "nomCompletImmeuble" : {
          "type" : "string"
        },
        "nomBaptiserImmeuble" : {
          "type" : "string"
        },
        "descriptionImmeuble" : {
          "type" : "string"
        },
        "numImmeuble" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrEtage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nbrePiecesDansImmeuble" : {
          "type" : "integer",
          "format" : "int32"
        },
        "idSite" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nomPropio" : {
          "type" : "string"
        },
        "prenomProprio" : {
          "type" : "string"
        },
        "garrage" : {
          "type" : "boolean"
        }
      }
    },
...
    "PaysDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "abrvPays" : {
          "type" : "string"
        },
        "nomPays" : {
          "type" : "string"
        }
      }
    },
...
    "PrestationAdditionnelReservationSaveOrrUpdate" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idReservation" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idServiceAdditionnelle" : {
          "type" : "integer",
          "format" : "int64"
        },
        "namePrestaion" : {
          "type" : "string"
        },
        "amountPrestation" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
    "PrestationSaveOrUpdateDto" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "number",
          "format" : "double"
        }
      }
    },
...
    "ReservationSaveOrUpdateDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "utilisateurRequestDto" : {
          "$ref" : "#/definitions/UtilisateurRequestDto"
        },
        "idAppartementdDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idUtilisateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "bienImmobilierOperation" : {
          "type" : "string"
        },
        "designationBail" : {
          "type" : "string"
        },
        "abrvCodeBail" : {
          "type" : "string"
        },
        "enCoursBail" : {
          "type" : "boolean"
        },
        "archiveBail" : {
          "type" : "boolean"
        },
        "montantCautionBail" : {
          "type" : "number",
          "format" : "double"
        },
        "nbreMoisCautionBail" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nouveauMontantLoyer" : {
          "type" : "number",
          "format" : "double"
        },
        "idBienImmobilier" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idLocataire" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeAbrvBienImmobilier" : {
          "type" : "string"
        },
        "pourcentageReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "montantReduction" : {
          "type" : "number",
          "format" : "double"
        },
        "soldReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "montantPaye" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbreAdulte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "montantReservation" : {
          "type" : "number",
          "format" : "double"
        },
        "nmbrEnfant" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "RoleRequestDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleName" : {
          "type" : "string"
        },
        "descriptionRole" : {
          "type" : "string"
        }
      }
    },
...
    "UtilisateurRequestDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idAgence" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idCreateur" : {
          "type" : "integer",
          "format" : "int64"
        },
        "utilisateurIdApp" : {
          "type" : "string"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "mobile" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "string",
          "format" : "date"
        },
        "lieuNaissance" : {
          "type" : "string"
        },
        "typePieceIdentite" : {
          "type" : "string"
        },
        "numeroPieceIdentite" : {
          "type" : "string"
        },
        "dateDebutPiece" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFinPiece" : {
          "type" : "string",
          "format" : "date"
        },
        "nationalite" : {
          "type" : "string"
        },
        "genre" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "profileImageUrl" : {
          "type" : "string"
        },
        "lastLoginDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastLoginDateDisplay" : {
          "type" : "string",
          "format" : "date-time"
        },
        "joinDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "roleUsed" : {
          "type" : "string"
        },
        "authorities" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "agenceDto" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roleRequestDto" : {
          "$ref" : "#/definitions/RoleRequestDto"
        },
        "userCreate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idEtablissement" : {
          "type" : "integer",
          "format" : "int64"
        },
        "active" : {
          "type" : "boolean"
        },
        "activated" : {
          "type" : "boolean"
        },
        "nonLocked" : {
          "type" : "boolean"
        }
      }
    },
...
  }
}